@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,10 +27,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  speech_to_text: ^2.3.0
-  highlight_text: ^0.7.2
-  avatar_glow: ^1.2.0
+  cupertino_icons: ^1.0.6
+  speech_to_text: ^6.4.1
+  highlight_text: ^1.7.0
+  avatar_glow: ^2.0.2
 
 dev_dependencies:
   flutter_test:
